updated models/debugging routes
