+++ app/models/__init__.py
@@ -1,6 +1,2 @@
 from .db import db
-from .user import User
-from .server import  Server
-from .channel import Channel
-from .message import Message
-from .friend import Friend
+from .models import Server, User, Channel, Friend, Message
